backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
google-genai
sentence-transformers
faiss-cpu
pandas
pydantic
python-dotenv
groq
